requirements.txt
streamlit>=1.35.0
pyserial>=3.5
pandas>=2.0
pytest>=8.0


